# src/objects/api/v2/openapi.yaml
openapi: 3.0.3
info:
  title: Objects API
  version: 2.6.0
  description: |
    An API to manage Objects.

    # Introduction

    An OBJECT is of a certain OBJECTTYPE (defined in the Objecttypes API). An
    OBJECT has a few core attributes that every OBJECT (technically a RECORD,
    see below) has, although these attributes can sometimes be empty. They are
    attributes like `geometry` and some administrative attributes. The data that
    describes the actual object is stored in the `data` attribute and follows
    the JSON schema as given by the OBJECTTYPE.

    ## Validation

    When an OBJECT is created or changed the `OBJECT.type` attribute refers to the
    matching OBJECTTYPE in the Objecttypes API. The RECORD always indicates which
    OBJECTTYPE-VERSION is used, shown in the `RECORD.typeVersion` attribute.

    Using these 2 attributes, the appropriate JSON schema is retrieved from the
    Objecttypes API and the OBJECT data is validated against this JSON schema.

    ## History

    Each OBJECT has 1 or more RECORDs. A RECORD contains the data of an OBJECT
    at a certain time. An OBJECT can have multiple RECORDS that describe the
    history of that OBJECT. Changes to an OBJECT actually create a new RECORD
    under the OBJECT and leaves the old RECORD as is.

    ### Material and formal history

    History can be seen from 2 perspectives: formal and material history. The
    formal history describes the history as it should be (stored in the
...
    attribute).
...
paths:
...
  /objects/{uuid}:
    get:
      operationId: object_read
      description: Retrieve a single OBJECT and its actual RECORD. The actual record
        is defined as if the query parameter `date=<today>` was given.
      parameters:
      - in: header
        name: Accept-Crs
        schema:
          type: string
          enum:
          - EPSG:4326
        description: 'The desired ''Coordinate Reference System'' (CRS) of the response
          data. According to the GeoJSON spec, WGS84 is the default (EPSG: 4326 is
          the same as WGS84).'
      - in: query
        name: fields
        schema:
          type: string
        description: 'Comma-separated fields, which should be displayed in the response.
          For example: ''url, uuid, record__geometry''.'
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
          description: Unique identifier (UUID4)
        required: true
      tags:
      - objects
      security:
      - tokenAuth: []
      responses:
        '200':
          headers:
            Content-Crs:
              schema:
                type: string
                enum:
                - EPSG:4326
              description: 'The ''Coordinate Reference System'' (CRS) of the request
                data. According to the GeoJSON spec, WGS84 is the default (EPSG: 4326
                is the same as WGS84).'
            API-version:
              schema:
                type: string
              description: 'Geeft een specifieke API-versie aan in de context van
                een specifieke aanroep. Voorbeeld: 1.2.1.'
            X-Unauthorized-Fields:
              schema:
                type: string
              description: 'List of fields that are not allowed to display if the
                field-based authorization is turned on. The value has the following
                format: `objectType1:fieldA,fieldB; objectType2:fieldC,fieldD`'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Object'
          description: OK
    put:
      operationId: object_update
      description: Update the OBJECT by creating a new RECORD with the updates values.
      parameters:
      - in: header
        name: Accept-Crs
        schema:
          type: string
          enum:
          - EPSG:4326
        description: 'The desired ''Coordinate Reference System'' (CRS) of the response
          data. According to the GeoJSON spec, WGS84 is the default (EPSG: 4326 is
          the same as WGS84).'
      - in: header
        name: Content-Crs
        schema:
          type: string
          enum:
          - EPSG:4326
        description: 'The ''Coordinate Reference System'' (CRS) of the request data.
          According to the GeoJSON spec, WGS84 is the default (EPSG: 4326 is the same
          as WGS84).'
        required: true
      - in: header
        name: Content-Type
        schema:
          type: string
          enum:
          - application/json
        description: Content type of the request body.
        required: true
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
          description: Unique identifier (UUID4)
        required: true
      tags:
      - objects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Object'
        required: true
      security:
      - tokenAuth: []
      responses:
        '200':
          headers:
            Content-Crs:
              schema:
                type: string
                enum:
                - EPSG:4326
              description: 'The ''Coordinate Reference System'' (CRS) of the request
                data. According to the GeoJSON spec, WGS84 is the default (EPSG: 4326
                is the same as WGS84).'
            API-version:
              schema:
                type: string
              description: 'Geeft een specifieke API-versie aan in de context van
                een specifieke aanroep. Voorbeeld: 1.2.1.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Object'
          description: OK
    patch:
      operationId: object_partial_update
      description: Update the OBJECT by creating a new RECORD with the updates values.
        The provided `record.data` value will be merged recursively with the existing
        record data.
      parameters:
      - in: header
        name: Accept-Crs
        schema:
          type: string
          enum:
          - EPSG:4326
        description: 'The desired ''Coordinate Reference System'' (CRS) of the response
          data. According to the GeoJSON spec, WGS84 is the default (EPSG: 4326 is
          the same as WGS84).'
      - in: header
        name: Content-Crs
        schema:
          type: string
          enum:
          - EPSG:4326
        description: 'The ''Coordinate Reference System'' (CRS) of the request data.
          According to the GeoJSON spec, WGS84 is the default (EPSG: 4326 is the same
          as WGS84).'
        required: true
      - in: header
        name: Content-Type
        schema:
          type: string
          enum:
          - application/json
        description: Content type of the request body.
        required: true
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
          description: Unique identifier (UUID4)
        required: true
      tags:
      - objects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedObject'
      security:
      - tokenAuth: []
      responses:
        '200':
          headers:
            Content-Crs:
              schema:
                type: string
                enum:
                - EPSG:4326
              description: 'The ''Coordinate Reference System'' (CRS) of the request
                data. According to the GeoJSON spec, WGS84 is the default (EPSG: 4326
                is the same as WGS84).'
            API-version:
              schema:
                type: string
              description: 'Geeft een specifieke API-versie aan in de context van
                een specifieke aanroep. Voorbeeld: 1.2.1.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Object'
          description: OK
    delete:
      operationId: object_delete
      description: Delete an OBJECT and all RECORDs belonging to it.
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
          description: Unique identifier (UUID4)
        required: true
      - in: query
        name: zaak
        schema:
          type: string
          format: uri
        description: '**Experimental** When destructing for archiving a ztc.ZAAK pass
          in the ZAAK URL that is being destroyed. If this OBJECT''s current RECORD
          has other ZAAK references, then object destruction is cancelled and only
          this ZAAK URL is removed from the RECORD''s references.'
      tags:
      - objects
      security:
      - tokenAuth: []
      responses:
        '200':
          headers:
            Content-Crs:
              schema:
                type: string
                enum:
                - EPSG:4326
              description: 'The ''Coordinate Reference System'' (CRS) of the request
                data. According to the GeoJSON spec, WGS84 is the default (EPSG: 4326
                is the same as WGS84).'
            API-version:
              schema:
                type: string
              description: 'Geeft een specifieke API-versie aan in de context van
                een specifieke aanroep. Voorbeeld: 1.2.1.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BehoudenResponse'
          description: OBJECT kept because it has multiple ZAKEN. Specified ZAAK parameter
            removed from RECORD.
        '204':
          headers:
            API-version:
              schema:
                type: string
              description: 'Geeft een specifieke API-versie aan in de context van
                een specifieke aanroep. Voorbeeld: 1.2.1.'
          description: OBJECT and all its RECORDs deleted.
...
  /objecttypes/{objecttype_uuid}/versions:
    get:
      operationId: objecttypeversion_list
      description: Retrieve all versions of an OBJECTTYPE
      parameters:
      - in: path
        name: objecttype_uuid
        schema:
          type: string
          format: uuid
        description: Unique identifier (UUID4)
        required: true
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: pageSize
        required: false
        in: query
        description: 'Het aantal resultaten terug te geven per pagina. (default: 100).'
        schema:
          type: integer
      tags:
      - objecttypes
      security:
      - tokenAuth: []
      responses:
        '200':
          headers:
            API-version:
              schema:
                type: string
              description: 'Geeft een specifieke API-versie aan in de context van
                een specifieke aanroep. Voorbeeld: 1.2.1.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedObjectTypeVersionList'
          description: OK
    post:
      operationId: objecttypeversion_create
      description: Create an OBJECTTYPE with the given version.
      parameters:
      - in: header
        name: Content-Type
        schema:
          type: string
          enum:
          - application/json
        description: Content type of the request body.
        required: true
      - in: path
        name: objecttype_uuid
        schema:
          type: string
          format: uuid
        description: Unique identifier (UUID4)
        required: true
      tags:
      - objecttypes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObjectTypeVersion'
      security:
      - tokenAuth: []
      responses:
        '201':
          headers:
            API-version:
              schema:
                type: string
              description: 'Geeft een specifieke API-versie aan in de context van
                een specifieke aanroep. Voorbeeld: 1.2.1.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectTypeVersion'
          description: Created
...
components:
  schemas:
    BehoudenResponse:
      type: object
      properties:
        behouden:
          type: array
          items:
            type: string
            format: uri
          description: Kept OBJECT URLs
      required:
      - behouden
...
    Object:
      type: object
      description: |-
        this mixin allows selecting fields for serializer in the query param
        It also supports nested fields.
      properties:
        url:
          type: string
          format: uri
          readOnly: true
          minLength: 1
          maxLength: 1000
          description: URL reference to this object. This is the unique identification
            and location of this object.
        uuid:
          type: string
          format: uuid
          description: Unique identifier (UUID4)
        type:
          type: string
          format: uri
          minLength: 1
          maxLength: 1000
          description: Url reference to OBJECTTYPE
        record:
          allOf:
          - $ref: '#/components/schemas/ObjectRecord'
          description: State of the OBJECT at a certain time
      required:
      - record
      - type
...
    PaginatedObjectTypeVersionList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/ObjectTypeVersion'
...
    PatchedObject:
      type: object
      description: |-
        this mixin allows selecting fields for serializer in the query param
        It also supports nested fields.
      properties:
        url:
          type: string
          format: uri
          readOnly: true
          minLength: 1
          maxLength: 1000
          description: URL reference to this object. This is the unique identification
            and location of this object.
        uuid:
          type: string
          format: uuid
          description: Unique identifier (UUID4)
        type:
          type: string
          format: uri
          minLength: 1
          maxLength: 1000
          description: Url reference to OBJECTTYPE
        record:
          allOf:
          - $ref: '#/components/schemas/ObjectRecord'
          description: State of the OBJECT at a certain time
...
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"
servers:
...